+++ webpack.config.js
@@ -1,10 +1,11 @@
 const path = require('path')
+const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 const TerserPlugin = require('terser-webpack-plugin')
 
 module.exports = {
   entry: {
-    main: path.resolve(__dirname, 'src/assets/main/src/scripts/main.js'),
-    preview: path.resolve(__dirname, 'src/assets/preview/src/scripts/preview.js')
+    main: path.resolve(__dirname, 'src/assets/main/src/scripts/main.ts'),
+    preview: path.resolve(__dirname, 'src/assets/preview/src/scripts/preview.ts')
   },
   output: {
     path: path.resolve(__dirname, 'src/assets/'),
@@ -15,6 +16,9 @@ module.exports = {
     craft: 'Craft',
     garnish: 'Garnish'
   },
+  resolve: {
+    extensions: ['.ts', '.tsx']
+  },
   optimization: {
     minimize: true,
     minimizer: [new TerserPlugin()]
@@ -22,25 +26,34 @@ module.exports = {
   module: {
     rules: [
       {
-        test: /\.jsx?$/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader',
-          options: { presets: ['@babel/preset-env'] }
-        }
+        use: ['ts-loader'],
+        include: [
+          path.resolve(__dirname, 'src/assets/main/src'),
+          path.resolve(__dirname, 'src/assets/preview/src')
+        ],
+        test: /\.tsx?$/
       },
       {
-        test: /\.scss$/,
-        use: [
-          { loader: 'style-loader' },
-          { loader: 'css-loader' },
-          { loader: 'sass-loader' }
-        ]
+        use: ['source-map-loader'],
+        enforce: 'pre',
+        test: /\.js$/
+      },
+      {
+        use: [MiniCssExtractPlugin.loader, 'css-loader'],
+        test: /\.css$/
+      },
+      {
+        use: [MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader'],
+        test: /\.scss$/
       },
       {
         test: /\.(png|svg)$/,
         type: 'asset/inline'
       }
     ]
-  }
+  },
+  devtool: 'source-map',
+  plugins: [new MiniCssExtractPlugin({
+    filename: '[name]/dist/styles/[name].css'
+  })]
 }